# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GoJetter API",
        "description": "API documentation for the GoJetter Laravel backend.",
        "contact": {
            "email": "support@gojetter.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/gojetter",
            "description": "GoJetter API Server"
        }
    ],
    "paths": {
...
        "/api/logout": {
            "post": {
                "tags": [
                    "Customer Auth"
                ],
                "summary": "Logout the authenticated customer user",
                "operationId": "a575315a0c547a3759c22fc7a2852083",
                "responses": {
                    "200": {
                        "description": "Successful Logout",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Message": {
                                            "type": "string",
                                            "example": "User logged out successfully."
                                        },
                                        "Success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "Data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": []
                                        },
                                        "Status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/resend-email-otp": {
            "post": {
                "tags": [
                    "Customer Auth"
                ],
                "summary": "Resend email OTP to an existing user",
                "operationId": "f10d09fa35facc7f074fe9ff887ff2cf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user7@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP resent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Message": {
                                            "type": "string",
                                            "example": "OTP sent successfully."
                                        },
                                        "Success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "Data": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "Status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error (email required or not found)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The email field is required."
                                        },
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The selected email is invalid."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/forget-password": {
            "post": {
                "tags": [
                    "Customer Auth"
                ],
                "summary": "Send OTP to email for password reset",
                "operationId": "d810648d68d0622b4216c9e8cc283f6e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Registered user email",
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Message": {
                                            "type": "string",
                                            "example": "OTP sent successfully."
                                        },
                                        "Success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "Data": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "Status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "Errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "BearerToken": {
                "type": "apiKey",
                "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\"",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "BearerToken": []
        }
    ]
}